refactor: modularize config manager
