+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -195,6 +195,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 680 to 200 lines by extracting core, collector, analyzer, and config modules. Main orchestrator now imports these modules while preserving functionality.
 
+### MODERATE-037: Config Manager Modularization ✅
+- **File**: `src/core/config/config_manager.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-25
+- **Summary**: Split the monolithic configuration manager into core, parser, and validator modules with a lightweight orchestrator.
+
 
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
