+++ src/core/config/config_manager.py
@@ -1,7 +1,7 @@
 """Configuration manager orchestrator.
 
-This thin facade wires together the loader, validator and core
-implementation.  It exposes the same public interface as the original
+This thin facade wires together the parser, validator and core
+implementation. It exposes the same public interface as the original
 monolithic module but delegates the heavy lifting to specialised
 components created during initialization.
 """
@@ -9,13 +9,13 @@ components created during initialization.
 from __future__ import annotations
 
 from .config_manager_core import ConfigManagerCore
-from .config_manager_loader import ConfigLoader
-from .config_manager_validator import ConfigValidator
-from .config_manager_config import ConfigEntry, ConfigMetadata
+from .config_parser import ConfigParser
+from .config_validator import ConfigValidator
+from ..config_manager_config import ConfigEntry, ConfigMetadata
 
 __all__ = [
     "ConfigManager",
-    "ConfigLoader",
+    "ConfigParser",
     "ConfigValidator",
     "ConfigEntry",
     "ConfigMetadata",
@@ -27,4 +27,4 @@ class ConfigManager(ConfigManagerCore):
 
     def __init__(self, config_dir: str | None = None):
         config_dir = config_dir or "config"
-        super().__init__(config_dir, ConfigLoader, ConfigValidator)
+        super().__init__(config_dir, ConfigParser, ConfigValidator)
