+++ src/core/performance_validation_system_refactored.py
@@ -6,7 +6,7 @@ import logging
 from typing import Any, Dict, Optional
 
 from .agent_manager import AgentManager
-from .config_manager import ConfigManager
+from .config.config_manager import ConfigManager
 from .assignment_engine import ContractManager
 from .workflow import (
     WorkflowOrchestrator as AdvancedWorkflowEngine,
